[SMALLFIX] Converted all the int type of fileId to long type

+++ clients/unshaded/src/main/java/tachyon/client/UfsUtils.java
@@ -161,7 +161,7 @@ public class UfsUtils {
           LOG.info("File " + tfsPath + " already exists in Tachyon.");
           continue;
         }
-        int fileId = tfs.createFile(tfsPath, ufsPath);
+        long fileId = tfs.createFile(tfsPath, ufsPath);
         if (fileId == -1) {
           LOG.warn("Failed to create tachyon file: " + tfsPath);
         } else {
